Move PageCountContext to the pages directory
